app/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
pydantic
google-cloud-storage
requests